// src/model/dto/consultaUsuario.dto.ts
export class ConsultaUsuarioDto{    
    
    IDPERSONA: number;
    COMUNA: string;
    TIPODOCUMENTO: string;
    RUT: string;
    CORREOELECTRONICO: string;
    NOMBRE: string;
    APELLIDOPATERNO: string;
    APELLIDOMATERNO: string;
    FECHANACIMIENTO: string;
    CELULAR: string;
    PLANTA: string;
    NACIONALIDAD: string;
    SEXO: string;
    ESTADO_PERSONA: string;
    IDUSUARIO: number;
    FECHACREACION: string;
    ESTADO_USUARIO: string;
    NOMBRE_JEFATURA: string;
    NOMBRE_GERENTE: string;    
    
}